wip: first pass PR changes
